manacher: use u64

+++ src/manacher.rs
@@ -1,6 +1,6 @@
 #[snippet = "manacher"]
 #[doc = "manacher[i] is the palindrome radius at i"]
-pub fn manacher(s: &[usize]) -> Vec<usize> {
+pub fn manacher(s: &[u64]) -> Vec<usize> {
     let mut r = vec![0; s.len()];
     let mut i = 0;
     let mut j = 0;
